+++ src/loader/file/level/level.cpp
@@ -220,7 +220,7 @@ void Level::convertTexture(ByteTexture& tex, Palette& pal, DWordTexture& dst)
     }
   }
 
-  dst.md5 = util::md5(&tex.pixels[0][0], 256u * 256u);
+  dst.md5 = util::md5(&tex.pixels[0][0], gsl::narrow_cast<size_t>(256u * 256u));
 }
 
 void Level::convertTexture(WordTexture& tex, DWordTexture& dst)
